+++ oem/ibm/libpldmresponder/file_io_type_cert.cpp
@@ -104,14 +104,14 @@ int CertHandler::write(const char* buffer, uint32_t offset, uint32_t& length,
     }
 
     auto fd = std::get<0>(it->second);
-    int rc = lseek(fd, offset, SEEK_SET);
+    int rc = static_cast<int>(lseek(fd, offset, SEEK_SET));
     if (rc == -1)
     {
         error("CertHandler::write:lseek failed, ERROR={ERR}, OFFSET={OFFSET}",
               "ERR", errno, "OFFSET", offset);
         return PLDM_ERROR;
     }
-    rc = ::write(fd, buffer, length);
+    rc = static_cast<int>(::write(fd, buffer, length));
     if (rc == -1)
     {
         error(
